Moved most of hdf5 creation script to methods
